Add MainViewModel
